requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==3.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
oauthlib==3.2.2
openai==1.64.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
sniffio==1.3.1
tqdm==4.67.1
tweepy==4.15.0
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3